+++ src/util/graphic.ts
@@ -660,7 +660,7 @@ export const getComboBBox = (children: ComboTree[], graph: IGraph): BBox => {
 
   children.forEach(child => {
     const childItem = graph.findById(child.id);
-    if (!childItem.isVisible()) return; // ignore hidden children
+    if (!childItem || !childItem.isVisible()) return; // ignore hidden children
     childItem.set('bboxCanvasCache', undefined);
     const childBBox = childItem.getCanvasBBox();
     if (childBBox.x && comboBBox.minX > childBBox.minX) comboBBox.minX = childBBox.minX;